+++ Sasquatch/SasquatchSwift/AppDelegate.swift
@@ -316,7 +316,7 @@ extension AppDelegate: DistributeDelegate {
     return false
   }
   
-  func onNoReleaseAvailable(_ distribute: Distribute) {
+  func distributeNoReleaseAvailable(_ distribute: Distribute) {
     let alert = UIAlertController(title: nil, message: "No updates available", preferredStyle: .alert)
     self.window?.rootViewController?.present(alert, animated: true)
   }