+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ### App Center Distribute
 
-* **[Feature]** Add `onNoReleaseAvailable` callback to Distribute delegate.
+* **[Feature]** Add `distributeNoReleaseAvailable` callback to Distribute delegate.
 * **[Feature]** Add the Distribute module to Swift Package Manager. Please note that due to additional resources bundle, it requires Swift 5.3 (Xcode 12 and higher).
 
 ___